+++ src/repository/mapper/ProjectMapper.php
@@ -22,6 +22,7 @@ class ProjectMapper
         $projectEntity->setTechno($data['techno'] ?? null);
         $projectEntity->setSubscriptionName($data['subscriptionName'] ?? null);
         $projectEntity->setMdmWorkloadVersion($data['mdmWorkloadVersion'] ?? null);
+        $projectEntity->setPathLivenessProbe($data['pathLivenessProbe'] ?? null);
         $projectEntity->setWebUrl($data['webUrl'] ?? '');
         $projectEntity->setArchived($data['archived'] ?? false);
         $projectEntity->setUrlHealthCheck($data['urlHealthCheck'] ?? []);
@@ -49,6 +50,7 @@ class ProjectMapper
         $project->setTechno($data['techno'] ?? null);
         $project->setSubscriptionName($data['subscriptionName'] ?? null);
         $project->setMdmWorkloadVersion($data['mdmWorkloadVersion'] ?? null);
+        $project->setPathLivenessProbe($data['pathLivenessProbe'] ?? null);
         $project->setWebUrl($data['webUrl'] ?? '');
         $project->setArchived($data['archived'] ?? false);
         $project->setUrlHealthCheck($data['urlHealthCheck'] ?? []);
@@ -76,6 +78,7 @@ class ProjectMapper
         $project->setTechno($entity->getTechno());
         $project->setSubscriptionName($entity->getSubscriptionName());
         $project->setMdmWorkloadVersion($entity->getMdmWorkloadVersion());
+        $project->setPathLivenessProbe($entity->getPathLivenessProbe());
         $project->setWebUrl($entity->getWebUrl());
         $project->setArchived($entity->isArchived());
         $project->setUrlHealthCheck($entity->getUrlHealthCheck());
@@ -110,6 +113,7 @@ class ProjectMapper
             $project->getUrlFronts(),
             $project->getUrlPubsubs(),
             $project->getMdmWorkloadVersion(),
+            $project->getPathLivenessProbe(),
             $project->getUrlsRundeck(),
             $project->getUrlsDeploymentGcp()
         );
@@ -129,6 +133,7 @@ class ProjectMapper
             'techno' => $projectEntity->getTechno(),
             'subscriptionName' => $projectEntity->getSubscriptionName(),
             'mdmWorkloadVersion' => $projectEntity->getMdmWorkloadVersion(),
+            'pathLivenessProbe' => $projectEntity->getPathLivenessProbe(),
             'webUrl' => $projectEntity->getWebUrl(),
             'archived' => $projectEntity->isArchived(),
             'urlHealthCheck' => $projectEntity->getUrlHealthCheck(),