+++ tests/service/RepositoryServiceTest.php
@@ -1,24 +1,28 @@
 <?php
 declare(strict_types=1);
 
+namespace App\tests\service;
 
 use App\exception\TechnicalException;
 use App\service\RepositoryService;
-use App\tests\service\AbstractServiceCase;
+use App\tests\AbstractTestCase;
 use org\bovigo\vfs\vfsStream;
 use org\bovigo\vfs\vfsStreamDirectory;
 
 
-class RepositoryServiceTest extends AbstractServiceCase
+final class RepositoryServiceTest extends AbstractTestCase
 {
     private vfsStreamDirectory $root;
 
     final protected function setUp(): void
     {
-        AbstractServiceCase::setUp();
+        AbstractTestCase::setUp();
         $this->root = vfsStream::setup();
     }
 
+    /**
+     * @throws TechnicalException
+     */
     final public function testSaveAndReadFile(): void
     {
         $service = new RepositoryService(vfsStream::url('root'), self::$loggerFactory);
@@ -31,6 +35,9 @@ class RepositoryServiceTest extends AbstractServiceCase
         $this->assertEquals(json_encode($data), $this->root->getChild($filename)->getContent());
     }
 
+    /**
+     * @throws TechnicalException
+     */
     final public function testReadFile(): void
     {
         $service = new RepositoryService(vfsStream::url('root'), self::$loggerFactory);