+++ CHANGELOG.md
@@ -5,11 +5,17 @@ Historique de tous les changements notables du projet dev-console
 Format basé sur [Keep a Changelog](https://keepachangelog.com/fr/1.0.0/).
 
 
+## [1.2.0] - 17/06/2026
+
+### Added
+
+- Interface rundeck
+
 ## [1.1.0] - 17/06/2026
 
 ### Added
 
-- Interface monitoring 
+- Interface monitoring
 
 ## [1.0.0] - 22/05/2026